pyproject.toml
[build-system]
requires = ["hatchling>=1.24"]
build-backend = "hatchling.build"

[project]
name = "cce"
version = "0.1.0"
description = "Code Context Engine — 4-layer indexer for Django/FastAPI/React with LangGraph agent runtime"
readme = "README.md"
requires-python = ">=3.12"
license = { text = "MIT" }
authors = [{ name = "Journies" }]

dependencies = [
    # --- core / config / cli ---
    "pydantic>=2.8",
    "pydantic-settings>=2.4",
    "typer>=0.12",
    "rich>=13.7",
    "python-ulid>=3.0",
    "pathspec>=0.12",

    # --- http server ---
    "fastapi>=0.115",
    "uvicorn[standard]>=0.30",
    "sse-starlette>=2.1",

    # --- agent runtime ---
    "langchain>=0.3",
    "langchain-core>=0.3",
    "langgraph>=0.2",
    "langgraph-checkpoint-sqlite>=2.0",

    # --- parsing (Phase 3+) ---
    "tree-sitter>=0.23",
    "tree-sitter-python>=0.23",
    "tree-sitter-javascript>=0.23",
    "tree-sitter-typescript>=0.23",
    "jedi>=0.19",

    # --- stores ---
    "qdrant-client>=1.12",

    # --- watcher (Phase 10) ---
    "watchdog>=5.0",
]

[project.optional-dependencies]
embeddings = [
    # nomic-embed-code default backend
    "torch>=2.4",
    "transformers>=4.45",
    "accelerate>=0.34",
    "sentencepiece>=0.2",
]
graph = [
    "kuzu>=0.7",
]
agents = [
    "langchain-openai>=0.2",
    "langchain-anthropic>=0.3",
    "langchain-community>=0.3",
]
dev = [
    "pytest>=8.3",
    "pytest-asyncio>=0.24",
    "httpx>=0.27",
    "ruff>=0.6",
    "mypy>=1.11",
]

[project.scripts]
cce = "cce.cli:app"

[tool.hatch.build.targets.wheel]
packages = ["cce"]

[tool.ruff]
line-length = 110
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]
ignore = ["E501"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = "-ra -q"

[tool.mypy]
python_version = "3.12"
strict = true
ignore_missing_imports = true
